# services/ontology-definition-service/api/openapi.yaml
...
openapi: "3.1.0"
info:
  title: OpenFoundry ontology-definition-service
  version: "1.0.0"
  description: |
    Foundry-style schema-of-types control plane: object types, link
    types, properties, action types (in `ontology-actions-service`),
    interfaces, shared property types, and object-type groups.

    Every mutation here writes to a transactional outbox in the same
    Postgres transaction (ADR-0022). Debezium tails the WAL and
    re-emits the events on the canonical Kafka topics:

      - ontology.object_type.changed.v1
      - ontology.link_type.changed.v1
      - ontology.property.changed.v1
      - ontology.object_type_group.changed.v1
      - ontology.interface.changed.v1
      - ontology.shared_property_type.changed.v1

    Authentication: every `/api/v1/ontology*` route requires a Bearer
    JWT. The token's `sub` claim is recorded as `actor_id` / `owner_id`
    in the audit log and in every outbox event.
servers:
  - url: https://gateway.openfoundry.example.com
    description: Gateway prefix; routes mounted on `/api/v1/ontology`.
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/ontology/object-types/{id}/properties/{propertyID}:
    parameters:
      - { in: path, name: id,         required: true, schema: { type: string, format: uuid } }
      - { in: path, name: propertyID, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [properties]
      summary: Read one property.
      responses:
        "200": { $ref: "#/components/responses/Property" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [properties]
      summary: Partially update one property.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdatePropertyRequest" }
      responses:
        "200": { $ref: "#/components/responses/Property" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [properties]
      summary: Delete one property.
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/ontology/links/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [link-types]
      summary: Read one link type.
      responses:
        "200": { $ref: "#/components/responses/LinkType" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [link-types]
      summary: Partially update a link type.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLinkTypeRequest" }
      responses:
        "200": { $ref: "#/components/responses/LinkType" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [link-types]
      summary: Delete a link type.
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:     { description: Resource not found. }
...
    LinkType:          { description: Link type record.,          content: { application/json: { schema: { $ref: "#/components/schemas/LinkType" } } } }
    Property:          { description: Property record.,           content: { application/json: { schema: { $ref: "#/components/schemas/Property" } } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
...
    Property:
      type: object
      properties:
        id:                { $ref: "#/components/schemas/Uuid" }
        object_type_id:    { $ref: "#/components/schemas/Uuid" }
        name:              { type: string }
        display_name:      { type: string }
        description:       { type: string }
        property_type:     { type: string, description: "Base type token (string, integer, float, boolean, timestamp, uuid, geopoint, geoshape, …)." }
        required:          { type: boolean }
        unique_constraint: { type: boolean }
        time_dependent:    { type: boolean }
        default_value:     {}
        validation_rules:  {}
        version:           { type: integer }
...
    UpdatePropertyRequest:
      type: object
      properties:
        display_name:      { type: string }
        description:       { type: string }
        property_type:     { type: string }
        required:          { type: boolean }
        unique_constraint: { type: boolean }
        time_dependent:    { type: boolean }
        default_value:     {}
        validation_rules:  {}
    LinkType:
      type: object
      properties:
        id:              { $ref: "#/components/schemas/Uuid" }
        name:            { type: string }
        display_name:    { type: string }
        description:     { type: string }
        source_type_id:  { $ref: "#/components/schemas/Uuid" }
        target_type_id:  { $ref: "#/components/schemas/Uuid" }
        cardinality:     { type: string, enum: [one_to_one, one_to_many, many_to_one, many_to_many] }
        label:           { type: string }
        reverse_label:   { type: string }
        visibility:      { type: string, enum: [normal, hidden] }
        owner_id:        { $ref: "#/components/schemas/Uuid" }
        created_at:      { $ref: "#/components/schemas/Timestamp" }
        updated_at:      { $ref: "#/components/schemas/Timestamp" }
        version:         { type: integer }
...
    UpdateLinkTypeRequest:
      type: object
      properties:
        display_name:  { type: string }
        description:   { type: string }
        cardinality:   { type: string, enum: [one_to_one, one_to_many, many_to_one, many_to_many] }
        label:         { type: string }
        reverse_label: { type: string }
        visibility:    { type: string, enum: [normal, hidden] }
...